requirements.txt
pandas
sqlalchemy
psycopg2-binary
streamlit
python-dotenv 